fix: Cast value output to float32 in evaluate() before numpy conversion
evaluate() was missing the .float() cast that evaluate_batch() and
evaluate_leaves() already had, causing a BFloat16 TypeError when
running on CUDA with autocast enabled.

+++ mcts/evaluator.py
@@ -185,7 +185,7 @@ class NNEvaluator:
         policy_probs = torch.softmax(policy_logits.float(), dim=-1).squeeze(0).cpu().numpy()
 
         # Value: already has tanh applied in the model, un-rotate to canonical
-        values_rotated = value_output.squeeze(0).cpu().numpy()
+        values_rotated = value_output.float().squeeze(0).cpu().numpy()
         canonical_values = unrotate_values(values_rotated, active_player)
 
         return policy_probs, canonical_values, mask_np
